requirements.txt
flask
flask-sqlalchemy
python-dotenv
google-generativeai
psycopg2-binary
gunicorn